explode CardStyles into a new directory
