+++ src/cardStyles/BackgroundType.ts
@@ -1,0 +1,8 @@
+/**
+ * Background style types for card backs
+ */
+export enum BackgroundType {
+  SOLID = 'solid',
+  GRADIENT = 'gradient',
+  IMAGE = 'image',
+}
