+++ src/cardStyles/CardTheme.ts
@@ -1,0 +1,8 @@
+/**
+ * Card theme types
+ */
+export enum CardTheme {
+  CLASSIC = 'classic',
+  MODERN = 'modern',
+  MINIMAL = 'minimal'
+}
