+++ src/cardStyles/GradientConfig.ts
@@ -1,0 +1,8 @@
+/**
+ * Gradient configuration
+ */
+export interface GradientConfig {
+  type: 'linear' | 'radial';
+  angle?: number; // for linear gradients (in degrees)
+  colors: string[]; // array of color stops
+}
