+++ src/cardStyles/SymbolStyle.ts
@@ -1,0 +1,12 @@
+/**
+ * Symbol style configuration
+ */
+
+export interface SymbolStyle {
+  fontSize: number;
+  fontWeight: string;
+  heartColor: string;
+  diamondColor: string;
+  clubColor: string;
+  spadeColor: string;
+}
