+++ src/cardStyles/ThemeConfig.ts
@@ -1,0 +1,11 @@
+import { CardTheme } from '.';
+
+/**
+ * Theme configuration interface
+ */
+
+export interface ThemeConfig {
+  name: CardTheme;
+  displayName: string;
+  description: string;
+}
