+++ src/game/index.ts
@@ -1,16 +1,5 @@
-export { Card, Suit, Rank } from './Card';
+export { Card, Rank, Suit } from './Card';
 export { Deck } from './Deck';
-export { Player } from './Player';
 export { Game } from './Game';
-export { 
-  BackgroundType, 
-  CARD_STYLE_PRESETS, 
-  DEFAULT_CARD_STYLE 
-} from './CardStyles';
-export type { 
-  GradientConfig, 
-  CardBackStyle, 
-  SymbolStyle, 
-  CardFrontStyle, 
-  CardStyleConfig 
-} from './CardStyles';
+export { Player } from './Player';
+
