+++ src/settings/SettingsManager.ts
@@ -1,44 +1,4 @@
-import type { CardStyleConfig } from '../game/CardStyles';
-import { DEFAULT_CARD_STYLE, CARD_STYLE_PRESETS } from '../game/CardStyles';
-
-/**
- * Card theme types
- */
-export enum CardTheme {
-  CLASSIC = 'classic',
-  MODERN = 'modern',
-  MINIMAL = 'minimal',
-}
-
-/**
- * Theme configuration interface
- */
-export interface ThemeConfig {
-  name: CardTheme;
-  displayName: string;
-  description: string;
-}
-
-/**
- * Available card themes
- */
-export const CARD_THEMES: ThemeConfig[] = [
-  {
-    name: CardTheme.CLASSIC,
-    displayName: 'Classic',
-    description: 'Traditional playing card style',
-  },
-  {
-    name: CardTheme.MODERN,
-    displayName: 'Modern',
-    description: 'Contemporary design with bold colors',
-  },
-  {
-    name: CardTheme.MINIMAL,
-    displayName: 'Minimal',
-    description: 'Clean and simple design',
-  },
-];
+import { CARD_STYLE_PRESETS, CardTheme, DEFAULT_CARD_STYLE, CardStyleConfig } from '../cardStyles';
 
 /**
  * Settings manager for the application
@@ -120,13 +80,13 @@ export class SettingsManager {
       'theme-modern',
       'theme-minimal'
     );
-    
+
     // Add current theme class
     document.body.classList.add(`theme-${this.currentTheme}`);
-    
+
     // Dispatch custom event for theme change
-    window.dispatchEvent(new CustomEvent('themechange', { 
-      detail: { theme: this.currentTheme } 
+    window.dispatchEvent(new CustomEvent('themechange', {
+      detail: { theme: this.currentTheme }
     }));
   }
 
@@ -208,8 +168,8 @@ export class SettingsManager {
     }
 
     // Dispatch custom event for card style change
-    window.dispatchEvent(new CustomEvent('cardstylechange', { 
-      detail: { cardStyle: this.currentCardStyle } 
+    window.dispatchEvent(new CustomEvent('cardstylechange', {
+      detail: { cardStyle: this.currentCardStyle }
     }));
   }
 }
