+++ src/settings/index.ts
@@ -1,2 +1,1 @@
-export { SettingsManager, CardTheme, CARD_THEMES } from './SettingsManager';
-export type { ThemeConfig } from './SettingsManager';
+export * from './SettingsManager';
